small stylish changes
